requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git@master
python-multipart>=0.0.6
apscheduler>=3.10.0
pydantic-settings>=2.0.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
